+++ src/main/java/insane96mcp/stamina/stamina/StaminaHandler.java
@@ -6,35 +6,62 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 
 public class StaminaHandler {
-    public static float getMaxStamina(Player player) {
-        int maxStamina = Mth.ceil(player.getHealth()) * StaminaFeature.staminaPerHalfHeart;
-        float ratio = player.getHealth() / player.getMaxHealth();
+    /**
+     * Returns the maximum possible stamina (when the player is fully healed)
+     */
+    public static float getMaxPossibleStamina(Player player) {
+        float maxPossibleStamina = Mth.ceil(player.getMaxHealth()) * StaminaFeature.staminaPerHalfHeart;
         if (player.getAttribute(StaminaFeature.BONUS_STAMINA_ATTRIBUTE.get()) != null)
-            maxStamina += (int) (player.getAttributeValue(StaminaFeature.BONUS_STAMINA_ATTRIBUTE.get()) * ratio);
+            maxPossibleStamina += (float) player.getAttributeValue(StaminaFeature.BONUS_STAMINA_ATTRIBUTE.get());
         if (StaminaFeature.staminaPerLevelOfVigourEnchantment > 0) {
             int enchLvl = EnchantmentHelper.getEnchantmentLevel(StaminaFeature.VIGOUR.get(), player);
             if (enchLvl > 0)
-                maxStamina += (int) (StaminaFeature.staminaPerLevelOfVigourEnchantment * enchLvl * ratio);
+                maxPossibleStamina += StaminaFeature.staminaPerLevelOfVigourEnchantment * enchLvl;
         }
         for (MobEffectInstance instance : player.getActiveEffects()) {
             if (instance.getEffect() instanceof IStaminaModifier staminaModifier)
-                maxStamina += (int) (staminaModifier.bonusMaxStamina(instance.getAmplifier()) * ratio);
+                maxPossibleStamina += staminaModifier.bonusMaxStamina(instance.getAmplifier());
         }
-        return maxStamina;
+        return maxPossibleStamina;
     }
 
-    public static float getStaminaPerHalfHeart(Player player) {
+    /**
+     * Returns the current max stamina
+     */
+    public static float getMaxStamina(Player player) {
+        float ratio = player.getHealth() / player.getMaxHealth();
+        return getMaxPossibleStamina(player) * ratio;
+    }
+
+    /**
+     * Returns the current max stamina
+     */
+    public static float getMaxStamina(Player player, float maxPossibleStamina) {
+        float ratio = player.getHealth() / player.getMaxHealth();
+        return maxPossibleStamina * ratio;
+    }
+
+    /*public static float getStaminaPerHalfHeart(Player player) {
         return getMaxStamina(player) / player.getHealth();
-    }
+    }*/
 
+    /**
+     * Returns the current stamina
+     */
     public static float getStamina(Player player) {
         return player.getPersistentData().getFloat(StaminaFeature.STAMINA);
     }
 
+    /**
+     * Returns true if the player is locked from sprinting
+     */
     public static boolean isStaminaLocked(Player player) {
         return player.getPersistentData().getBoolean(StaminaFeature.STAMINA_LOCKED);
     }
 
+    /**
+     * Returns true if the player can sprint
+     */
     public static boolean canSprint(Player player) {
         return !isStaminaLocked(player);
     }