+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Config options have been sorted better
 * Fixed stamina not synced when switching dimensions
+* Fixed stamina not locked below 2.5 max health
 
 ## 1.0.2
 * Added a config option to disable swimming (split from sprinting)